Switch back to using int instead of size_t.

+++ src/instruction.c
@@ -59,7 +59,7 @@ void cpu_draw(struct cpu* cpu) {
     for (int x = 0; x < 8; x++) {
       uint8_t pixel = cpu->memory[cpu->i + y];
       if (pixel & (0x80 >> x)) {
-        size_t index =
+        int index =
             (cpu->v[cpu->opcode.x] + x) % SCREEN_WIDTH +
             ((cpu->v[cpu->opcode.y] + y) % SCREEN_HEIGHT) * SCREEN_WIDTH;
         if (cpu->pixels[index] == ON_COLOR) {
@@ -76,7 +76,7 @@ void cpu_draw(struct cpu* cpu) {
 
 void cpu_wait_key_press(struct cpu* cpu) {
   cpu->pc -= 2;
-  for (size_t i = 0; i < 16; i++) {
+  for (int i = 0; i < 16; i++) {
     if (SDL_GetKeyboardState(NULL)[key_map[i]]) {
       cpu->v[cpu->opcode.x] = i;
       cpu->pc += 2;
